chore: add page (#7)

+++ tzkt-client/src/main/kotlin/com/rarible/tzkt/model/Part.kt
@@ -1,3 +1,3 @@
 package com.rarible.tzkt.model
 
-data class Part(val address: String, val share: Long)
+data class Part(val address: String, val share: Int)
